backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
Pillow>=10.2.0
pydantic>=2.5.0
